be rfc compliant...
RFC 822 *only* supports four digits for the year in order to have a compliant
header.
RFC 850 supports two digits for the year (y2k_incompliant :) so use that when
y2k_compliance is turned off.

In either case, both of these formats require that GMT be appended to the header,
so we do this as well


git-svn-id: http://svn.php.net/repository/pear/pear-core/pear/pear-core/trunk@65024 c90b9560-bf6c-de11-be94-00142212c4b1