add <usesrole>/<usestask> validation.
usesrole/task is designed to provide a graceful error
message for a package.xml that uses a custom role
or task.  If the package that provides the custom role/task
is not yet installed, the error message will instruct the
user to install the package and which package to install.


git-svn-id: http://svn.php.net/repository/pear/pear-core/pear/pear-core/trunk@182345 c90b9560-bf6c-de11-be94-00142212c4b1