Fixed bug #18001	PEAR/RunTest coverage fails [dufuz]
In rev 297621 Sebastian B. added a fix to make sure the Namespace declaration stays at the top. As a result the coverage code was no longer included in the output .php file - It has been added again.
There is also a PHP Notice fix in the affected code in case the first line is unset (if it's <?php or <?)