Add basic framework to verify a GnuPG signature
Will correctly print an error message and exit immediately if a bad
signature is produced.  Ignores the signature and continues installation
if the public key is not available.  Continues installation on a good
signature.

This is very basic checking, and UI may not be totally desirable - gpg
output is printed directly to console - but it is a start.