Fix tests that broke with PHP7 but were not really broken, only wrong
PHP 7 uses a different order when calling the comparison method in
usort(), so we get a different order if there are elements which are
equal. This is very apparent when computing order for packages that are
going to be uninstalled and that do not depend on each other.

Since we support both PHP 5 and 7 :) we cannot rely on a determined
order and have to adjust the tests.