Add proper HTTP proxy support through the CONNECT verb
At the moment, we handle proxy requests for HTTPS URLs (almost) the same
as regular non-encrypted HTTP requests. This means that we send a line
similar to this to the proxy:
    GET https://pear.horde.org HTTP/1.1

Many proxies do not handle proxy requests when they contain a HTTP URL
(for example, Squid) and fail e.g. with a 501 Not Implemented error.

The proper way to make these requests is through the CONNECT verb.

This not only solves the issues with proxies that do not handle the
aforementioned GET requests with HTTPS URLs, but also increses security,
as we can do the SSL certificate validation by ourselves and do not have
to trust the proxy.

Additionally, we no longer speak HTTPS with the proxy in case we request
a HTTPS URL as it is simply no longer needed.