Remove -n option from PECL script for shared extensions
When PHP is using shared extensions:

./configure --with-openssl=shared --enable-xml=shared

The -n option in PHP command line (in the PECL shell script) removes them
from being enabled and PECL script won't work ok anymore. This goes mainly
for the xml and openssl PHP extensions which are needed to install PECL
extensions via the https protocol and to parse XML files.