Fixed extension loaded check for pecl binaries
This check is used to make sure we don't overwrite binary files that are
already in use, as doing so will cause a segmentation fault. This is reported
among others in https://bugs.xdebug.org/view.php?id=2167, but I have ran into
this in many occasions.

However, this check never worked, since its introduction 19 years ago, as the
`return` was missing, and the suffix as created by `basename` does not include
the `.`.